shell/flow: split Emscripten code into platform files
Move all EM_JS bindings, IDBFS persistence, start gate, touch
controls, and GL context setup from flow.c into flow_emscripten.c
with no-op stubs in flow_generic.c, following the existing
log_linux.c / log_windows.c / log_emscripten.c pattern.

Extract video pixel format constants into video_emscripten.h /
video_generic.h headers selected by video_platform.h, removing
ifdefs from fmv.c.

Make Shell_PersistConfigToIDBFS and Shell_PersistSavesToIDBFS
unconditionally available (no-op on generic), removing ifdefs
from config/common.c and savegame/common.c.

Remove redundant whole-file ifdef guard from log_emscripten.c.
