+++ src/meson.build
@@ -796,7 +796,9 @@ common_sources += desktop_sources
 if is_emscripten
   common_sources += ['trx/core/log_emscripten.c']
   common_sources += ['trx/gl/gl_webgl_compat.c']
+  common_sources += ['trx/game/shell/flow_emscripten.c']
 else
+  common_sources += ['trx/game/shell/flow_generic.c']
   if dep_backtrace.found() and host_machine.system() == 'linux'
     common_sources += ['trx/core/log_linux.c']
   elif host_machine.system() == 'windows'
