+++ src/trx/av/video_emscripten.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <trx/gl/gl_webgl_compat.h>
+
+#include <libavutil/pixfmt.h>
+
+#define VIDEO_PIX_FMT AV_PIX_FMT_RGBA
+#define VIDEO_TEX_FORMAT GL_RGBA
+#define VIDEO_TEX_TYPE GL_UNSIGNED_BYTE
