+++ src/trx/av/video_generic.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <GL/glew.h>
+
+#include <libavutil/pixfmt.h>
+
+#define VIDEO_PIX_FMT AV_PIX_FMT_BGRA
+#define VIDEO_TEX_FORMAT GL_BGRA
+#define VIDEO_TEX_TYPE GL_UNSIGNED_INT_8_8_8_8_REV
