+++ src/trx/game/fmv.c
@@ -2,6 +2,7 @@
 
 #include <trx/av/audio.h>
 #include <trx/av/video.h>
+#include <trx/av/video_platform.h>
 #include <trx/config.h>
 #include <trx/core/filesystem.h>
 #include <trx/core/log.h>
@@ -35,13 +36,8 @@ static OUTPUT_QUAD_SURFACE_DESC M_MakeSurfaceDesc(
         .width = width,
         .height = height,
         .bit_count = 32,
-#ifdef EMSCRIPTEN_BUILD
-        .tex_format = GL_RGBA,
-        .tex_type = GL_UNSIGNED_BYTE,
-#else
-        .tex_format = GL_BGRA,
-        .tex_type = GL_UNSIGNED_INT_8_8_8_8_REV,
-#endif
+        .tex_format = VIDEO_TEX_FORMAT,
+        .tex_type = VIDEO_TEX_TYPE,
         .uv = {
             { .u = 0.0f, .v = 0.0f },
             { .u = 1.0f, .v = 0.0f },
@@ -180,11 +176,7 @@ static bool M_Play(const char *const file_name)
         Video_SetSurfaceSize(
             video, Viewport_GetWidth(VIEWPORT_GAME),
             Viewport_GetHeight(VIEWPORT_GAME));
-#ifdef EMSCRIPTEN_BUILD
-        Video_SetSurfacePixelFormat(video, AV_PIX_FMT_RGBA);
-#else
-        Video_SetSurfacePixelFormat(video, AV_PIX_FMT_BGRA);
-#endif
+        Video_SetSurfacePixelFormat(video, VIDEO_PIX_FMT);
 
         Video_PumpEvents(video);
